Inger ingerasul meu

+++ src/health_management/Student.java
@@ -12,7 +12,7 @@ public class Student extends Utilizator {
 	
 	public void completareFormular(int idFormular) {
 		ArrayList<Integer> raspunsuri = FormularRegistry.get(idFormular).completareFormular();
-		float scor = FormularRegistry.get(idFormular).calculareScor(raspunsuri);
+		int scor = FormularRegistry.get(idFormular).calculareScor(raspunsuri);
 		
 		formulareCompletate.add(new FormularCompletat(idFormular, raspunsuri, scor));
 	}
